template,echo: add ability to use per-handler middlewares
This is essentially the same thing as #259, but for echo.

I could not find a way to access the OAuth scope of the route from the normal
middlewares, as the value gets set by ServerInterfaceWrapper. With this change,
it's now possible to add a middleware to check that the scope is valid prior
to running the actual handler.

This commit introduces a new Middlewares field to ServerInterfaceWrapper,
which is a slice of echo.MiddlewareFunc, that gets executed in order
after the route parameters get decoded. It also adds a new registration
function, RegisterHandlersWithOptions, to allow users to specify these
middlewares.